feat: Complete functionality for dragover and gdagend events
Expected behaviour of moving a card from the tableau removes it from the
origin pile and adds it to the destination pile was correctly
implemented. In addition drag end events are handled unhiding the card
that was being dragged. This fixes the bug that a card would dissappear
if a drag ended without being dropped on a valid pile